Use PUT methods instead of manually detecting if new or update
Kong supports resource idempotency by calling the `PUT` method on user
creation/update.

Kong will decide by itself whether to create or update the resource.

It's clearer and less code on our side. However we can only do this on
named resources (not possible on Plugins and Routes which are attached
to the `service` with an ID).